testing/web-platform/tests/workers/WorkerLocation_href.htm

<!DOCTYPE html>
<html>
<head>
<title> WorkerLocation href attribute </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var FileName = "./support/WorkerLocation.js?srch%20#hash";
    var href = window.location.href;
    var ExpectedResult = href.substring(0, href.lastIndexOf('/')) + FileName.substring(1);
    
    var t = async_test("Test Description: WorkerLocation href attribute must return the absolute URL that the WorkerLocation object represents.");
                  
    var worker = new Worker(FileName);

    worker.onmessage = t.step_func(function(evt)
    {
        assert_equals(evt.data.href, ExpectedResult);
        t.done();
    });
</script>
</body>
</html>
